se agrego barra styky y se agrego alert para eliminar datos

+++ src/Components/VocabularyListItem.jsx
@@ -36,7 +36,9 @@ const VocabularyListItem = ({ form, openEdit, deleteItem, color }) => {
   };
 
   const deleteRegistry = () => {
-    deleteItem(form.id);
+    if (confirm("Estas seguro de eliminar " + form.word)) {
+      deleteItem(form.id);
+    }
   };
   return (
     <>
